feat: Implement UUIDs for Primary Keys (#19)
* feat: Implement UUIDs for Primary Keys

This commit implements the use of UUIDs as primary keys across the project. The following changes were made:

- Migrations: Updated all relevant migrations to use `uuid` type for primary keys and foreign keys.
- Models: Modified model files to use UUIDs as primary keys, including setting `$primaryKey`, `$keyType`, `$incrementing`, and using the `HasUuids` trait.
- Repositories: change the methods and the queries to work with uuid instead of id.
- Requests: Change request validations and parameters to work with uuid instead of id.

This change enhances data integrity and prepares the project for distributed environments.

* fix(migrations): Remove duplicate uuid definition in failed_jobs migration

The `failed_jobs` migration file was creating the 'uuid' column twice, causing a database error. This commit corrects the migration by removing the duplicate definition, ensuring there is only one 'uuid' column defined as primary key.

* fix: Correct repository interface mismatches

This commit fixes the type mismatches between the UserRepository interface and implementation

* fix(tests): Update UsersTest to use UUIDs for user profiles and fix validation errors

This commit addresses test failures in UsersTest by making the following changes:

-   **User Profile Creation:** The tests for creating user profiles were failing because they were not providing a valid `user_id` (UUID). This commit modifies the tests to first create a user, then retrieve its UUID, and finally use that UUID when sending the request to create a user profile.
- Fixed user profile duplicate check.

These changes ensure that the tests correctly interact with the API now that we are using UUIDs instead of integer IDs.

* fix(user): Correct CreateUserProfileRequest validation for UUIDs

This commit updates the `CreateUserProfileRequest` to validate `user_id` as a UUID, fixing the "The selected user id is invalid." error and allowing user profiles to be created successfully.

* fix(tests): Correct UsersTest for UUIDs and user profile creation

This commit updates the UsersTest file to properly use UUIDs when creating user profiles. It ensures that the test creates a user first, retrieves the user's UUID, and then uses it when creating the user profile.

* refactor: update user_id to user_uuid for UUID consistency across user components

---------

Co-authored-by: Ajithkumar S <ajithkumar.sn@pitsolutions.com>